Set Rust edition to 2018 and minimum version to 1.31
Since this library does not use any advanced features and all of the
APIs come from really early-on Rust, set the minimum version to a really
long time ago. There does not appear to be a way to compile the examples
with this setting because of `clap`, but the library itself will work
just fine.